Remove mmap storage (move to examples)
